--- rets/interpreters/__init__.py
@@ -1,1 +1,0 @@
-from .search import SearchInterpreter
